+++ Editor/Utils/EditorUtility.cs
@@ -1,0 +1,15 @@
+﻿using System.Reflection;
+
+namespace EtienneEditor
+{
+    public sealed class EditorUtility
+    {
+        public static void ClearConsole()
+        {
+            Assembly assembly = Assembly.GetAssembly(typeof(UnityEditor.Editor));
+            System.Type type = assembly.GetType("UnityEditor.LogEntries");
+            MethodInfo method = type.GetMethod("Clear");
+            method.Invoke(new object(), null);
+        }
+    }
+}