+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Dates in European way
 
+## [0.1.2] - 7/03/2022
+### Fixed
+- ClearLog Build Bug
+
 ## [0.1.1] - 28/02/2022
 ### Fixed
 - UserPref Creation